Rename Data.Poly.Uni -> Data.Poly.Internal
